tests/wpt/web-platform-tests/web-share/share-url-invalid.https.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>WebShare Test: Share with an invalid URL</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
  </head>
  <body>
    <script>
      promise_test(async t => {
        const url = "http://example.com:65536";
        await test_driver.bless(
          "web share",
          () => {
            return promise_rejects_js(t, TypeError, navigator.share({ url }));
          },
          "share with an invalid URL"
        );
      }, "share() rejects when URL is invalid");
    </script>
  </body>
</html>
